+++ client/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.2.0
+
+* Added Simple Transactions
+* Fixed UnitsHelper.formatBest for -ve numbers
+* Added unmount detection for TransactionObject
+* Mam progress indicator updated
+* Removed live fields on Transaction Decoder
+* Linked obsolete tag on TransactionObject
+
 ## v0.1.0
 
 * Initial Release